shared: Move UUID generation functions to plain sol-util
They are generic enough that can be reused by other platforms without
problems.
